+++ CHANGELOG.md
@@ -2,10 +2,48 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
-## [36.8.1] 2025-01-28
+## [36.10.5] 2025-02-25
 ### Fixed
 - Fixed a display issue for answers, where the display field might
   not have data.
+
+## [36.10.4] 2025-02-24
+### Fixed
+- Fixed a bug that the repeating numbers are not displayed correctly 
+  on the repeating items when their data are loaded through FHIRPath
+  expressions.
+  
+## [36.10.3] 2025-02-07
+### Changed
+- Changed the meta.profile for STU3, R4 and R5 in the exported 
+  QuestionnaireResponse so that the standard FHIR profiles are used.
+
+## [36.10.2] 2025-02-03
+### Fixed
+- The "url" parameter was not URL-encoded when doing expansions.
+
+## [36.10.1] 2025-01-30
+### Fixed
+- Warning messages were being incorrectly generated for failing to load a value
+  set for an autocomplete field, for which the value set should not be and is
+  not loaded at form initialization time.
+- Improved the timing of the formReady event, so that it is sent after the form
+  has been rendered.
+
+## [36.10.0] 2025-01-30
+### Added
+- Support for rendering-xhtml on contained[valueSet].expansion.contains.display.
+
+## [36.9.0] 2025-01-30
+### Added
+- Support for rendering-xhtml on answerOption.valueCoding.display.
+
+## [36.8.2] 2025-01-17
+### Fixed
+- Fixed a bug that linkIds in help text items were lost in exported
+  Questionnaire.
+
+>>>>>>> pending-release
 ## [36.8.1] 2024-12-17
 ### Fixed
 - Fixed a bug that code.system was not set correctly when exporting